backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
lxml==5.3.0
defusedxml==0.7.1
itsdangerous==2.2.0
python-dotenv==1.0.1
